Delete colaborador

+++ app/Http/Controllers/Colaboradores/ColaboradoresController.php
@@ -9,6 +9,11 @@ use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
 use App\Modelos\User;
 use App\Modelos\Colaborador\DetalleColaborador;
+use App\Modelos\Oportunidad\ColaboradorOportunidad;
+use App\Modelos\Oportunidad\ArchivosOportunidadColaborador;
+use App\Modelos\Prospecto\ColaboradorProspecto;
+use App\Modelos\Prospecto\ArchivosProspectoColaborador;
+
 
 use DB;
 use Mail;
@@ -153,30 +158,82 @@ class ColaboradoresController extends Controller
     }
 
     public function deleteColaborador(Request $request){
-      $id_borrar = $request->borrar;
-      $id_asignar = $request->asignar;
+      $id_borrar = $request->id_borrar;
+      $id_asignar = $request->id_asignar;
 
-      $colaborador = User::where('id',$id_borrar)->first();
+      $colaborador_borrar = User::where('id',$id_borrar)->first();
+      $colaborador_asignar = User::where('id',$id_asignar)->first();
 
-      try{
-      if ($colaborador) {
-      DB::beginTransaction();
-      User::where('id',$id_borrar)->delete();
-      DB::commit();
-      return response()->json([
-          'message'=>'Successfully deleted',
-          'error'=>false,
-          'data'=> $id
-      ],200);
+      if ($colaborador_asignar) {
+        try{
+
+          DB::beginTransaction();
+
+          $oportunidades = ColaboradorOportunidad::where('id_colaborador',$id_borrar)->get();
+          foreach ($oportunidades as $oportunidad) {
+            $oportunidad->id_colaborador = $id_asignar;
+            $oportunidad->save();
+          }
+
+          $prospectos = ColaboradorProspecto::where('id_colaborador',$id_borrar)->get();
+          foreach ($prospectos as $prospecto) {
+            $prospecto->id_colaborador = $id_asignar;
+            $prospecto->save();
+          }
+
+          $archivos_prospecto = ArchivosProspectoColaborador::where('id_colaborador',$id_borrar)->get();
+          foreach ($archivos_prospecto as $archivo_prospecto) {
+            $archivo_prospecto->id_colaborador = $id_asignar;
+            $archivo_prospecto->save();
+          }
+
+          $archivos_oportunidad = ArchivosOportunidadColaborador::where('id_colaborador',$id_borrar)->get();
+          foreach ($archivos_oportunidad as $archivo_oportunidad) {
+            $archivo_oportunidad->id_colaborador = $id_asignar;
+            $archivo_oportunidad->save();
+          }
+
+          DB::commit();
+        }catch (Exception $e){
+
+          DB::rollBack();
+
+        }
       }
-    }catch (Exception $e){
-      DB::rollBack();
-      return response()->json([
-          'error'=>true,
-          'messages'=>'Colaborador no encontrado.',
 
+
+      if ($colaborador_borrar) {
+        try{
+
+            DB::beginTransaction();
+            User::where('id', $id_borrar)->delete();
+            DB::commit();
+
+            return response()->json([
+                'message'=>'Successfully deleted',
+                'error'=>false,
+                'data'=>$id_borrar
+            ],200);
+
+        }catch (Exception $e){
+
+          DB::rollBack();
+
+          return response()->json([
+            'message'=>'Something went grong',
+            'error'=>true,
+            'data'=>$id_borrar
+          ],400);
+
+        }
+      }
+
+      return response()->json([
+          'message'=>'Colaborador no encontrado',
+          'error'=>true,
+          'data'=>$id_borrar
       ],400);
-    }
+
     }
 
     public function transformColaboradorToJson(User $colaborador, DetalleColaborador $colaborador_ext){
